+++ src/app/achieve/cache/startsession_response.hpp
@@ -5,14 +5,14 @@
 namespace firelight::achievements {
     struct Unlock {
         int ID{};
-        int When{};
+        long long When{};
     };
 
     struct StartSessionResponse {
         bool Success{};
         std::vector<Unlock> HardcoreUnlocks;
         std::vector<Unlock> Unlocks;
-        int ServerNow{};
+        long long ServerNow{};
     };
 
     NLOHMANN_DEFINE_TYPE_NON_INTRUSIVE(Unlock, ID, When)
